% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{ari}
\alias{ari}
\title{Computes Adjusted Rand Index taking into account count of filtered out
singletons. Most of source code directly lifted from mclust package.}
\usage{
ari(f, c)
}
\arguments{
\item{f}{number of filtered out singletons}

\item{c}{corset & trinity filtered clustering data frame}
}
\value{
adjusted Rand Index
}
\description{
Computes Adjusted Rand Index taking into account count of filtered out
singletons. Most of source code directly lifted from mclust package.
}
