% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_unique.R
\name{best_unique}
\alias{best_unique}
\title{Best unique value from a column in a (BLAST) dataframe.}
\usage{
best_unique(df, col)
}
\arguments{
\item{df}{Data frame}

\item{col}{column to take unique results from}
}
\value{
Data frame filtered to only unique values of A with lowest evalue & highest % identity
}
\description{
Takes the unique result from column col with the lowest evalue then highest percent_identity. If there are still
multiple results from column col after the first one is selected.
Intended to select most significant BLAST results.
}
\examples{
df <- data.frame(col1=c('a','a', 'a'), evalue=c(1,2,1), percent_identity=c(94,99,94))
df2 <- best_unique(df,'col1')
}
