% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size_distribution.R
\name{cluster_size_distribution}
\alias{cluster_size_distribution}
\title{Returns cluster size distribution}
\usage{
cluster_size_distribution(clustering)
}
\arguments{
\item{clustering}{vector with gene names/IDs}
}
\value{
data frame with columns size (=cluster size) and freq (=frequency of that cluster)
}
\description{
Returns cluster size distribution
}
\examples{
# gene vector is based off of Trinity annotation but can be any format as long as it is gene names/IDs
# for example Corset annotation would be of format Cluster-1.0, Cluster-2.0, etc.
genes <- c("DN1_c0_g1","DN1_c0_g1","DN1_c0_g2","DN2_c0_g1")
cluster_size_distribution(genes)
# output is data frame with 2 clusters of size 1 and 1 cluster of size 2
}
