% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Computes cluster stats.}
\usage{
cluster_stats(df)
}
\arguments{
\item{df}{data frame}
}
\value{
List with num_singleton, num_same_subset, num_same_Y_diff_X
}
\description{
Computes from a data frame with clusterings in column 1 and column 2:
1) number of singleton clusters, 2) number of pairs in the same clustering in columns 1 and 2
3) number of pairs with the same clustering in 2 and different clusterings in 1, and
4) number of pairs with the same clustering in 1 and different clusterings in 2.
}
