% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{correctly_clustered}
\alias{correctly_clustered}
\title{Given a data frame with two columns consisting of clusterings,
gives number of entries that have been correctly clustered, assuming
that the 2nd column is the true clustering.}
\usage{
correctly_clustered(df)
}
\arguments{
\item{df}{Data frame with 2 columns of clusterings}
}
\value{
Number of entries that have been correctly clustered
}
\description{
Given a data frame with two columns consisting of clusterings,
gives number of entries that have been correctly clustered, assuming
that the 2nd column is the true clustering.
}
\examples{
df <- data.frame(col1=c(1,1,2,2,2,2,3,4,5),col2=c(1,1,2,3,3,3,4,4,5))
cc <- correctly_clustered(df)
}
