% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{different_subset}
\alias{different_subset}
\title{Computes number of pairs in a matrix from `table` that
are in the different subsets.}
\usage{
different_subset(tab)
}
\arguments{
\item{tab}{Object from `table`.}
}
\value{
Number of pairs that are in different subsets.
}
\description{
Computes number of pairs in a matrix from `table` that
are in the different subsets.
}
\examples{
df <- data.frame(col1=c(1,1,2,2,2,2,3,4,5),col2=c(1,1,2,3,3,3,4,4,5))
tab <- table(df$col1, df$col2)
ds <- different_subset(tab)
}
