% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size_dist.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plots cluster sizes and counts. Plot does not come with title; user has to add their own.
User is also recommended to adjust breaks with scale_x_discrete as well.}
\usage{
plot_clusters(df)
}
\arguments{
\item{df}{Data frame with columns: size, freq, Method, ID}
}
\description{
Plots cluster sizes and counts. Plot does not come with title; user has to add their own.
User is also recommended to adjust breaks with scale_x_discrete as well.
}
