% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trees.R
\name{processTree}
\alias{processTree}
\title{Helper function to read in newick trees; will skip over a non-existent
tree rather than return an error}
\usage{
processTree(x)
}
\arguments{
\item{string}{Path to tree file}
}
\value{
tree_text - Character string representing nhx tree
}
\description{
Helper function to read in newick trees; will skip over a non-existent
tree rather than return an error
}
