% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trees.R
\name{read_trees}
\alias{read_trees}
\title{Reads all newick trees in a path and parses with treedata structure so that
we can annotate with things like subtree length and height.}
\usage{
read_trees(path, extension, cores)
}
\arguments{
\item{path}{path to the newick trees}

\item{extension}{extension for newick trees}

\item{cores}{number of cores for mclapply}
}
\value{
A list of treedata objects
}
\description{
Reads all newick trees in a path and parses with treedata structure so that
we can annotate with things like subtree length and height.
}
