% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values.R
\name{replace_values}
\alias{replace_values}
\title{Replaces values in one column of df1 using df2. If value doesn't
exist in df2 then will replace with NA. Primarily intended for use
in replacing transcript IDs with gene IDs.}
\usage{
replace_values(df1, df2, col_id)
}
\arguments{
\item{df1}{Data frame 1, with values to be replaced}

\item{df2}{Data frame 2, with replacement values. Should be only 2 columns.}

\item{col_id}{Column ID of df1 to be replaced.}
}
\value{
Data frame 1 but with values in col_id replaced.
}
\description{
Replaces values in one column of df1 using df2. If value doesn't
exist in df2 then will replace with NA. Primarily intended for use
in replacing transcript IDs with gene IDs.
}
\examples{
df1 <- data.frame(col1=c(1,2), col2=c('a','b'))
df2 <- data.frame(ID1=c('a','b'), ID2=c('c','c'))
rv <- replace_values(df1, df2, 2)
}
