% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{same_subset}
\alias{same_subset}
\title{Computes number of pairs in a matrix from `table` that
are in the same subset. Intended to identify transcripts
that have been mapped to the same gene clusters between
either two different methods or one method and the true clustering.}
\usage{
same_subset(tab)
}
\arguments{
\item{tab}{Object from `table`.}
}
\value{
Number of pairs that are in the same subset.
}
\description{
Note: gene clusters consisting of a single transcript will not be identified
in this way. Use `identify_singletons` in combination with this.
}
\examples{
df <- data.frame(col1=c(1,1,2,2,2,2,3,4,5),col2=c(1,1,2,3,3,3,4,4,5))
tab <- table(df$col1, df$col2)
ss <- same_subset(tab)
}
