% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctly_clustered.R
\name{same_Y_diff_X}
\alias{same_Y_diff_X}
\title{Computes number of pairs in a matrix from `table` that
are in the same subset in Y but different subsets in X.}
\usage{
same_Y_diff_X(tab)
}
\arguments{
\item{tab}{Object from `table`.}
}
\value{
Number of pairs that are in the same subset in X & different in Y.
}
\description{
Computes number of pairs in a matrix from `table` that
are in the same subset in Y but different subsets in X.
}
\examples{
df <- data.frame(col1=c(1,1,2,2,2,2,3,4,5),col2=c(1,1,2,3,3,3,4,4,5))
tab <- table(df$col1, df$col2)
ds <- same_Y_diff_X(tab)
}
