% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size_dist.R
\name{single_cluster_df}
\alias{single_cluster_df}
\title{Creates a single data frame from a list of cluster size distribution data frames
with Method column indicating method (examples: Trinity, Corset, treeinform) and
ID column indicating sample (examples: SRX5, melanogaster).}
\usage{
single_cluster_df(list_cluster_dfs, methods, ids)
}
\arguments{
\item{list_cluster_dfs}{List of cluster size distribution data frames}

\item{methods}{Vector of methods. Must be same length as list_cluster_dfs}

\item{ids}{Vector of IDs. Must be same length as list_cluster_dfs}
}
\value{
Data frame with columns: size, freq, Method, ID
}
\description{
Creates a single data frame from a list of cluster size distribution data frames
with Method column indicating method (examples: Trinity, Corset, treeinform) and
ID column indicating sample (examples: SRX5, melanogaster).
}
\examples{
df1 <- data.frame(size=c(1,2), freq=c(10,20))
df2 <- data.frame(size=c(1,2), freq=c(5,5))
methods <- c("method1", "method2")
ids <- c("id1", "id1")
single_cluster_df(list(df1,df2),methods,ids)
}
