% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trinity_split.R
\name{trinity_split}
\alias{trinity_split}
\title{Split up trinity IDs}
\usage{
trinity_split(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Df but with trinity column split into trinity and isoform
}
\description{
Split up trinity IDs
}
